Parser changes to turn statement clauses into operators
We are currently treating statement clauses (e.g `FILTER`, `ORDER BY`)
as rigid statement syntax (a-la SQL).  Other parts of the grammar allow
`FILTER` and `ORDER BY` without an explicit statement keyword such as in
shapes and function arguments.  Furthermore, clauses in statements like
`DELETE` are a bit awkward semantically as `DELETE Expr FILTER expr`
sort-of violates the top-down flow semantics.

To avoid these and make things consistent throughout the language, this
removes statement clauses and instead turns `FILTER` et all into regular
infix operators.  So, `Expr FILTER Expr` is an expression, and
corresponding statements are redefined as `SELECT Expr`, `UPDATE Expr
SET ...` and `DELETE Expr`.

This commit contains only the necessary grammar changes as a proof that
this is possible and is fully compatible with current syntax.