Fix parser span over SQL adapter (#8399)
Closes https://github.com/geldata/gel/issues/8390

- **remove usage of lineno**

libpg_query's `PgQueryError` contains field `lineno`, which we assumed
is the line of the query source where the error occurred. It is not. It
is line of the .c file that has produced the parser error. We don't ever
want to use this.

- **minor optimization of inflate_span**

Just something I noticed.

- **don't emit bogus positions**

When we didn't have line and col info, we used to default to 0. Which is
worse than nothing, so I removed it.

This also adds code for computing line&col, but I'm not yet calling it,
because that requires the original source, which I don't know how to
get.