Fix schema reflection losing track of changes made while `using future`
Setting `using future simple_scoping` (or `warn_old_scoping`) can
cause expressions to be recompiled. When globals and expression
aliases are recompiled, their underlying ObjectType may be deleted and
then restored.

Not much depends on that, but it will frequently show up in expression
refs.

There is a bug in futures (there since the beginning) where we don't
reflect the changes made when compiling the future out to the
in-database schema, which can cause bad problems. Here, the problem is
that refs can refer to ObjectTypes for globals that don't exist in the
database.

The forward fix is easy: do that.

Fixes #8431.