Support databases that have this mistake in them already
Handling the case where the mistake has already been made is more
annoying.

The approach we take is to filter out expr `refs` that don't actually
exist in the schema. This does not really fix the state, but will
unbreak doing a dump, at least.

I think the state could also be fixed by dropping the future and
recreating it.