Fix schema reflection losing track of changes made while `using future` (#8435)
Setting `using future simple_scoping` (or `warn_old_scoping`) can cause
expressions to be recompiled. When globals and expression aliases are
recompiled, their underlying ObjectType may be deleted and then
restored.

Not much depends on that, but it will frequently show up in expression
refs.

There is a bug in futures (there since the beginning) where we don't
reflect the changes made when compiling the future out to the
in-database
schema, which can cause bad problems. Here, the problem is that refs can
refer to ObjectTypes for globals that don't exist in the database.

The forward fix is easy: do that.

Fixes #8431.

---
Handling the case where the mistake has already been made is more
annoying.

The approach we take is to filter out expr `refs` that don't actually
exist
in the schema. This does not really fix the state, but will unbreak
doing a
dump, at least.

I think the state could also be fixed by dropping the future and
recreating
it.

---

The test I added passes with *either* of the other two commits applied.