Fix regression with UPDATEs on a complex path with rewrite
This was introduced as a regression in 6.x by #7265, which eliminated
the use of a separate standalone subject_path_id to represent the
object being updated in a rewrite.

The method used for picking the the subject_path_id used when
compiling the rewrites was unsound, though, and only worked in the
simple case.

Fix this by passing in the path_id of the update expression, so that
it always matches.

(If this doesn't make sense to people, we could revert #7265 instead.)

Fixes #8444