pg-ext: add more SQL func wrappers (#8488)
Adds wrappers for:

* `has_database_privilege`
* `has_any_column_privilege` (this fixes Metabase compatibility)

If the same table name exists in more than one schema, the
`has*_privilege()` functions that take an unqualified table
name are not yet looking into `search_path` as they are expected
to be, but rather choose a "random" one as the result. This is
not a problem yet, because Gel user is "superuser" only, it has
access to any of the tables with the same name, so the result
is inaccurately correct.

Also fixes the `pg_database` view.

Refs #5307, #5319
Fixes #8457