Add server side parameter conversions (#8554)
Allows functions to specify a set of conversions to parameters by the server. The converted parameters must be either a constant or a query parameter.

This is useful for implementing ai text search, which will apply a conversion to a text param and convert it into an embedding.

Example param conversion:
```
    create function my_to_str(val: str) -> str
    {
        using (val);
    };
    create function my_to_str(val: int64) -> str
    {
        set server_param_conversions := (
            '{"val": "cast_int64_to_str"}'
        );
        using sql expression;
    };
```

Given the query `select simple_to_str(123)`, returns `"123"`

Functions can set `server_param_conversions` as a json string of the form `dict[str, str | list[str]]`.

## Flow of information

 In function resolution during `polyres.find_callable`:
- the conversions are parsed
- parameters will be substituted for converted parameters
  - in the example, `val: int64` becomes `val~cast_int64_to_str: str`
- gather any `additional_info`, a string tuple which is passed to the server
- another version of the function will be matched

Extra information for the server:
- converted parameters are specified as coming after globals during `clauses.populate_argmap`
- Each query unit can get conversions specified during `_compile_ql_query`

Doing the actual conversion during `execute.execute`:
- Gather the conversions and the encoded arg in `args_ser.get_param_conversions`
- Decode the arg and apply the conversion in `execute._convert_parameters`
- Pass the converted params into `args_ser.recode_bind_args`

For scripts, the unit group has an additional `unit_converted_param_indexes` parameter which tracks which parameters apply to which unit. If multiple units have the same conversion, it will only be performed once.

## Adding more conversions

To add a new conversion, the following locations should be updated:
- `edgeql.compiler.polyres._check_server_arg_conversion`
  - `conversion_name`: name of conversion to set in schema
  - `converted_type`: the resulting type of the conversion
  - `additional_info`: a string tuple which can be passed to the server
  - `conversion_volatility`: the volatility of the conversion, included in the compiled function's volatility
- `server.protocol.execute._convert_parameters`
  - implement the conversion