Fix regression in debug dumping of spans
PR #8538 rationalized the names of fields in spans, but this broke its
implementation of `as_markup`, which wasn't updated.
This code isn't really typechecked, apparently, and the failure is expressed only as
```
Exception during serialization to markup: <builtins.AttributeError: 'Span' object has no attribute 'name'>
```
being spewed when errors are printed to the server log, so no tests
failed and the server still worked.