Fix computed columns over SQL adapter
When compiling expressions of computed pointers
for SQL adapter, we provide it with an external
rvars of the current table.

We manually populate path_namespace which can be
used by the pgcompiler. For link tables, we
provide column `source` and for object type tables
we should have provided `id`.

But we didn't, which only broke when compiling
computeds that contained backlinks.