Cleanup a hack for DML on inherited links over SQL adapter (#8582)
For the following query:

```sql
WITH d AS (
  INSERT INTO "Document" (title) VALUES ('Report') RETURNING id
)
INSERT INTO "Document.shared_with" (source, target, can_edit)
SELECT d.id, $1, TRUE FROM d
```
... we generate an EdgeQL:
```
select {
  dml_0 :=insert Document,
  dml_1 :=update Document { shared_with += ...}
}
```

But to validate that ids inserted into `shared_with` exists in
`Document`,
we have to join with `Document` prior to insertion. That is problematic,
because we cannot just use the concrete table, since it does not yet
contain the inserted `Document`.

Until now, we used a hack that just didn't join with the concrete table.

This PR makes it so we join with concrete table AND the overlay that
contains
objects created during current query.

And removes the hack.

---

I've found a problem with this hack when I was trying to add a followup
test
case for #8562. The test case is not working correctly yet.