Fix a cross-type conflict bug that can occur with rewrites (#8584)
We were not properly considering rewrites when analyzing DML
operations to see if we needed to insert explicit checks to avoid
conflicting on exclusive properties.

So first we update `_get_needed_ptrs` to chase down rewrites, also
explicitly consider rewritten pointers (using the new
`_get_rewritten_ptrs`) and then switch to *always* using the
fake_dml_set machinery when using conflict checks for inheritance
purposes, so that we always look at the actually inserted data instead
of what appears in the edgeql query. This required some hacking on the
backend also, but not too much.

This also allows volatile properties to be used in DML that requires
conflict
checks, which is nice, because that was always a weird lurking problem.