Fix more server_param_conversion breakage of patches+IPU testing
The DESCRIBE of functions that use server_param_conversions wouldn't
output the server_param_conversions field, since it was not marked as
allow_ddl_set. Mark it as allow_ddl_set but manually prohibit users
from setting it.

This is very much on the borderline of being worth supporting.

Also make schema_repair() use testmode (which it was trying but
failing to do), so that schemas that use testmode features still work.
Patches actually was working in CI, but not locally, since it didn't
use this.