Tweak parser recovery
Closes #8609

This PR decreases cost of `if` and `group` keywords to prevent them from
being suggested, since they are usually not what is missing.

Also, I've adjusted how parsers are recovered and added a special path
for parsers with custom errors: when we detect a custom error, we will
not try to fix any other error and just skip tokens until EOI. This will
have the effect of never producing an AST if there is a custom file, but
reporting just the custom error and no other errors.

It does add a guarantee that the custom error will never be discarded in
lieu of other recovery paths.