Fix ISE when with binding in dml function contains a set function/operator (#8634)
Set funcs/ops such as ?? compile their arguments in a new scope.

If such funcs/ops are used in a volatile binding in a dml function, a non-volatile bindings which is compiled from the volatile binding will not have direct access to the args iterator rvar.

Instead, they need access via the args' scope stmt.