WIP: Start invalidating evicted cache entries on other frontends
We just weren't ever doing that. There was a commented out
signal_sysevent, but that forced a full reread, which was too
expensive to work.

This still only partially works, though, because there is still a race
condition. We will need to handle the errors and retry, I think.

Progress on #8636.