Fix some object-returning functions when using `future warn_old_scoping`
The issue was that (due to some obscure interactions in the frontend
maybe worth looking at also), we were feeding the backend compiler IR
that just had a TypeRoot as the toplevel expression, not a SelectStmt.

This ought to work (or ought to be disallowed forcefully), but was
breaking.

The issue is that the rvar for a TypeRoot is a relation, not an rvar
for the ctx.rel already created. In that case, we need to include the
rvar into the toplevel query, or the query will be empty.

Fixes #8645.