Gel-ls: go to definition for schema files
The implementation is a bit lacking, because it does not use the
existing code we have for name resolution, but instead just finds
ObjectRef under the cursor and looks up that name in the schema.

It does also look for enclosing `module` declarations, so it can resolve
unqualified names, but that's it. I hope our actual name resolution is
not any more complicated than this. We don't have a way to specify mod
aliases, right? We do have some custom handling for `std::`, but we
don't support definitions in `std` (yet) anyway.

I did make sure that definition lookup does *not* recompile schema,
since that is slow on large schema. It even caches parsed SDL, so
pressing `ctrl` and hovering over many identifiers does not overwhelm
the server.

To make this work, I had to improve how we set spans on AST, so we never
propagate spans from parents into children, but instead use spans
produced by the parser.