Implement sys::approximate_count()
It takes a schema::ObjectType, so it is called like
`sys::approximate_count(introspect Card)`.
This is because the implementation works by querying the `reltuples`
field of `pg_class`, and it really only makes sense for getting a
full table. Having it take `anytype` and special casing would
create weird discontinuities where the simple case worked like this
and complex cases have a correctness or performance cliff.

Maybe we want to name it something different?

By default it includes all descendant types too; the ignore_subtypes
parameter changes that.

Fixes #4164.