Gel-ls: improve go to definiton of with binding
Support for "go to definition of with binding" in EdgeQL. I had this
partially implemented before, but not tested. So it was not actually
working and had problems when a with binding type was used in a shape,
which created another, derived subtype of it. That subtype does not have
a span set, so we cannot jump to it's definition.

My solution was to 'walk up the inheritance tree' for such types that
do not have a span and are derived views.

I've also added a few debugging tools, needed to get debug messages
over LSP (or written to disk).