Gel-ls: completion for schema files
Adds a native implementation of completion for schema files to gel-ls.

It works by parsing the current file with recovery, finding the node
under the cursor and determining the module context of that node.

If module is found, we suggest all types in that module as completion
items.

We also suggest module names, regardless if we are in a module or not.

This works well when completing link types or extending types.