Disable DML on instered objects that needed for SQL (#8713)
Followup for #8582

For DML, I needed updates of objects that were inserted in the same
query. To be exact, I needed only *updates of their links*, which
translate to update of the link table only.

But my fix also affected EdgeQL DML, so this PR disables that.

We could keep this feature, but it would only work when updates really
are on links and multi properties only. Which would be a weird design
choice, so I would prefer not to have this feature in the first place.