Use provided callback URL in OAuth code exchange (#8728)
We recently updated the auth server to allow proxying the OAuth flow
through the application server to allow customizing the behavior or
keeping the database on a private network. One crucial part of this
proxying is being able to specify the internal `redirect_uri` that is
registered at the Identity Provider and used as part of the code
exchange between the auth server and the identity provider. We were
correctly setting this value on the initial authorize call to the IdP,
but then hardcoding the old direct URL.

In order to pass the correct URL through to this point in the flow, we
need to encode it into the `state`, and then use _that_ value instead of
the hard-coded value.