Do better multiplicity inference in nested FOR expressions (#8754)
Currently we disable our `disjoint_union` tricks inside nested FOR
expressions (because we need to make sure expressions like
`for x in {1,2} union (for y in {3, 4} union y)`) are DUPLICATE.

This is unnecessarily strict, though, especially because if there is a
shape in between the two fors, whether the inner FOR loop is UNIQUE
can matter, like in:
```
FOR movie IN Movie SELECT movie {
  title,
  actors := (FOR person IN movie.actors SELECT person),
};
```

This is easy to fix by considering the inner iterator but making sure
disjoint_union doesn't leak.

Fixes #8748.