Drop the spaces after field names in Tuple names
I used copilot to fix the tests, and it only took
me about ten times longer than doing it manually.
Copilot also kept wanting to suggest that it did
well, when writing this commit message.

One downside of this simple approach is that it
will change the ids of tuple types across versions
and even after a dump/restore (since the ids are
associated with names in dumps). The alternate
approach would be to apply a regex to fix the
names up when computing displaynames.