Use provided callback URL in OAuth code exchange (#8801)
Manual backport of https://github.com/geldata/gel/pull/8728

The difficulty in automatically backporting this change is that the JWT
handling in the `master` branch has been moved significantly around,
making it difficult to cleanly apply this to the `release/6.x` branch.
Instead of trying to match the move, this represents the smallest
possible change: in essence, it adds the `redirect_uri` to the state
token using the current mechanism present in the 6.x branch. This avoids
needing a much larger refactor and backport of the JWT handling code,
while keeping the behavioral change.