Support graphql over the binary protocol
This is implemented by adding another language enum field, and then
using the same flow as edgeql and SQL.

We'll only expose the JSON output mode in clients for now, because the
binary one doesn't work right for queries that do things like
introspection or @include.

Handling things like introspection and @include is the only reason
that this PR isn't trivial: graphql introspection queries are
implemented by having the compiler call out to the graphql-core
implementation to execute the query, and then including the json
response in the output. Making this work when the query might depend
on a variable requires tracking "key variables" and including them in
the cache keys, and recompiling when they don't match.

We already did this in the graphql HTTP endpoint but now need to
support it in the main flow.

Two main caveats to follow-up on:
 * This first take doesn't support doing the query normalization to
   extract constants. I'll do that in a follow-up.
 * We don't support queries that have default values for parameters.
   This is because the gel bindings don't have a way to *not pass*
   a parameter. (They can pass None, but that's not the same thing!)
   Not sure the best thing to do here.

Mostly finishes #8831.
(I'll close that when we land this and open follow-ups.)