Fix casting from JSON in indexes/constraints
This was a regression introduced in 6.8 by #8804, which changed the
volatility of a handful of metaschema functions, including
jsonb_extract_scalar. jsonb_extract_scalar is used by basically every
cast *from* json.

The issue is that the cast functions now are Stable, which breaks
things. I introduced a mechanism in the original patch whereby the
trampoline functions would stay marked Immutable, but this only
mattered when the generated queries were calling metaschema functions
directly, whereas casts have wrapper functions.

Our casts, unlike functions and operators, don't set the volatility of the
function. (And indeed, this is why changing the underlying volatility
of jsonb_extract_scalar made things faster--it let us eliminate a volatility
mismatch.)

The fix is to make sure the trampoline function for casts that are
called in indexes has volatility set.

We'll need to do a patch for 6.10 to update the trampoline
volatilities, but in-place upgrades shouldn't need any intervention,
since trampolines get recreated there automatically.