Add administer rebuild_sql_introspection command
Gel server supports updates of the backend postgresql version, without
any modification to contents of the database. That is, gel-server can
be shutdown, PostgreSQL database upgraded and then gel-server can
connect to the upgraded database without any additional steps.

Except SQL introspection views need to be rebuilt. That's because they
depend on the PostgreSQL version.

So this PR introduces `administer rebuild_sql_introspection()` command,
that does just that. It is idempotent (it can be run multiple times)
and should be run any time the version of the backend databases changes.