Fix updating of two single links with linkprops at once
Mutating a single link with link properties works by compiling the
actual expression that defines the value into the link table update,
and then copying that value into the table itself by (ab/re)using the
rewrite machinery.

The rewrite machinery uses a "policy_ctx", which uses one of the input
tables as part of the overlay instead of an output table. The
"policy_ctx" has its overlay tables copied before being updated, but
it does the copy "just in time", which means that you can sometimes
get away with depending on things added to the context *after* the
policy_ctx was created.

Of course, this only worked *once*, after which the copy had been
made. As a result there was a bug where an 'except' overlay wasn't
being applied when populating a single link.

Be more explicit about all this by populating *both* overlay tables.

Also, it turns out that we do depend on incorporating unrelated
changes made to the overlay tables during compilation (like if an
INSERT is happening nested inside other DML), so we need to explicitly
reapply the policy overlays to the main one. (There was an existing bug
caused by this that this change made appear more easily.)

Possibly also fix some other policy related issues, since the code is
shared, but I haven't tracked it down too much.