Avoid duplicating a cached computed global in some unnecessary cases (#8955)
Currently we always produce two CTEs for a computed global that is
referenced in both an access policy and outside of an access policy
(since it may produce different results in those contexts).

Add a check for whether any access policies are actually referenced,
and merge the CTEs if they are not.

This is important since #8953 would otherwise cause duplication of
auth JWT validation. This may also remove duplication in some existing
user queries.