Fix some optimizations in simple_scoping mode
I turned on simple_scoping for codegen tests and discovered that a
bunch was broken.

Two fixes:
 * Tweak the cardinality inference so that the code to detect
   ONE/AT_LEAST_ONE arguments to optional parameters and remove the
   optional flag works when there is an additional wrapping select.
 * Add a whole new optimization analysis/transform to the
   post-compilation work in stmtctx that finds the selects we
   inserted to implement simple_scoping and removes them when
   possible.

Both of these changes are exactly the kind of thing I hate about the
scope tree and path factoring, and it is an unfortunate irony that
getting rid of the scope tree requires first making it gnarlier.

The best way out is always through, Robert Frost says.