Don't lose overlays when iterating over DML with FOR
Since we use a VisibleBindingRef to refer to a for binding,
get_dml_sources was not finding dml sources of for variables.

Fix this by tracking the DML sources of FOR bindings in a symbol
table.

The other alternative choice here would be to make VisibleBindingRef
maintain a reference to the Set of of the iterator expression, so that
analyses can be done without needing symbol tables. (We could still
hide it from debug output.)

Fixes #8993.