Make SET REQUIRED USING (...) work on link properties
The SET REQUIRED USING path always used the `produce_ctes=True` mode
of `_compile_conversion_expression`, which depends on using `id` as
the key. That doesn't work for us here, since there's no `id` for a
link property.

If we supported using ctid or source, target as the key (see #5050) we
could make it work, but until then, use the `produce_ctes=False` mode
for link properties.

This means DML won't work there, but whatever.