Fix modifying and dropping inherited AI indexes
The bug is that the index's `kwargs` aren't inherited in subtypes
(which is bad), though the kwargs are still used in determining
the internal name of the index.
But worse, subsubtypes have no exposure to these kwargs, and
so the kwargs aren't used in determining the name, either,
which prevents it from getting deleted properly.

Make sure to include kwargs.

Fixes #9033.