Fix modifying and dropping inherited AI indexes (#9041)
The bug is that the index's `kwargs` aren't inherited in subtypes
(which is bad), though the kwargs are still used in determining the
internal name of the index. But worse, subsubtypes have no exposure to
these kwargs, and so the kwargs aren't used in determining the name,
either, which prevents it from getting deleted properly.

Make sure to include kwargs. Added some documentation to try to avoid
missing this kind of thing.

Might be worth reviewing by commits -- I also had to commit some hacks
in order to get `schema_repair` to work.

I tested that manually, and will try to test it automatically when I
backport this.

Fixes #9033.