Fix /rag on subtypes
When an AlterAnnotationValue is performed (this is triggered by the
annotation copying for some weird reason), we were losing the @expr
linkprop in the user-visible schema (reflected using AS_LINK).
(For actual internal use, annotations are reflected using an actual
AnnotationValue class, so this didn't cause any problems with the
internal schema.)

Fix this by making sure to populate all fields on AlterObjectFragments
of AS_LINK reflected objects. This ensures that the cruddy linkprop
based reflection can get the job done.

This is one of those classic cases that shows up a ton in our compiler
where some weird code path is being exercised in some specific case for
obscure reasons, and we are faced with the options of:
 1. Make some change to avoid the code path.
 2. Do some hacky fix in the particular code path.
 3. Try to do a general fix.

I did #3 here, though I think maybe it would have been better to do #2
and add an assert somewhere that we don't add any new uses of
AS_LINK...

We'll need to run a fixup script in the backport.
I think this will do it:
```
update schema::AnnotationSubject set {
    annotations := (
        for a in .annotations__internal
        select schema::Annotation { @value := a.value }
        filter .name = a.name
    )
};
```