Fix some issues with overloading AI indexes
When the "effective" index changes, we need to rename the pg index, so
ANALYZE won't get confused.

We also need to do a `refresh` when the effective index changes due to
a delete index, like we do for creates, since that will invalidate the
vectors.

Fix this by adding the new refresh call and making refresh rename the
index.

In PR #2474, I yanked out all the rename dbops, and remarked:
> If something changes in the future and we need to generate a pgsql
> rename, we can pull the old dbops code out of git.

Fixes #9048.