Fix normalization of calls of overloaded funcs over SQL adapter
Closes #8928

When calling functions that take param of type "Any", such as
`json_build_object`, PostgreSQL must infer the type of the param
somehow. When we replace arg `'hello'` with `$1`, that does not work
anymore.

So `select json_build_object($1, TRUE)` was reporting that is cannot
infer the type of $1.

My solution was to to reuse solution we use for `ROW('hello')`, which is
to detect that query param has type UNKNOWN (which happens for strings),
and inject `::text` around the query param.

For more details see: #8086