Pass `-I` in all the places we invoke `sys.executable` as subprocess
I wasted a bunch of time today because I had installed some stuff in
`.local` and it was getting picked up in inittestdb's subprocess
invocation (via cluster.py), and the click version skew was breaking
stuff.

I shouldn't have done that, I guess, but we ought to pass `-I`
(isolated).