Add a `.?>link` operator to dereference link while suppressing "hidden by policy"
Users (and us) occasionally have problems dealing with "required link
hidden by access policy" errors. They can be worked around often by
using backlinks and an explicit join, but that that is kind of
horrible and seems to sometimes be slow.

Issue #8522 proposes adding an intrinsic function to mask the error,
but I really hate that. I don't think that a simple function call
ought to be impacting the semantics of the expression contained
within, and I think it's really unclear what its behavior would be in
more complex cases.

Instead, add an operator, `Obj.?>link` that does an *optional*
dereference of `link`.  The result is always optional, and it
suppresses any "hidden by policy" error.

I'm happy to argue about syntax if anyone has something they like
more. I also considered `.?` but that was worse.

Closes #8522, but not in the way proposed there.